+++ .gitignore
@@ -13,4 +13,4 @@ examples/MNIST
 examples/multipart_serialised.eqx
 .python-version
 .DS_Store
-
+.ruff_cache
